refactor: creat auth file
