+++ backend/matcha/__init__.py
@@ -10,6 +10,7 @@ from flask_jwt_extended import (
 
 from matcha import app as approute
 from matcha import users
+from matcha import auth
 
 
 def create_app():
@@ -25,5 +26,6 @@ def create_app():
 
     app.register_blueprint(approute.bp)
     app.register_blueprint(users.bp)
+    app.register_blueprint(auth.bp)
 
     return app
